python-stats-service/requirements.txt
fastapi
uvicorn
python-dotenv
elevenlabs
google-generativeai
pymongo
motor
python-multipart